change the main looping mechanism
With the previous looping mechanism, it was not possible to properly count the
photons at levels other than TOA, because the last photons of each thread were
cut.
Now we use a new looping mechanism which avoids cutting the photons. NBLOOP
has a new meaning, it means the target number of photons for each kernel run.
Thus the kernel will run NBPHOTON/NBLOOP times.
NBLOOP will be renamed eventually, currently we keep it so for testing purposes
(compatibility with python scripts).
Looping now works like so:
- each thread is initially active and increments a global counter of active threads
- when the target number of photons for this thread is reached, all threads
  finish their photons. When a thread finished its last photon, it becomes
  inactive and decrements the global counter of active threads.
- when all threads are inactive, the kernel ends.