Add new locations, SURF0M and SURF0P
There was no mean to know whether the photon was reaching the surface from the
air or water side.
The test was on vz, which is not right.
Thus we introduce new states SURF0M and SURF0P, which determine if the surface
has been reached from above or brelow.