Bugfix in move_pp
This bug is revealed by the introduction of new locations SURF0M and SURF0P.
The test on tau for a photon to go to the water should be > 0, not >= 0.
If RAND is equal to zero, tau is equal to zero, thus a downward photon can be
reset to surface even though it crossed the surface already.
-> In previous scheme, there was no distinction between SURF0M and SURF0P,
thus the downward photon was just crossing the surface a second time (which
was wrong but unnoticeable).
-> With the new scheme, the downward photon would be set to SURF0M, leading to
an infinite while loop when searching for a compatible wave slope in the next
surface interaction.