Make a Smartg class. Remove smartg_thr.
We now use a class Smartg.
Compilation and kernel loading is done in the constructor.
Running is done in the calc(...) method.
This is done to speed up smartg loops, where a non-negligible time was
spent loading the kernel, even with cached compilation results (~0.3s).

Another benefit is the logical separation between compilation and
runtime.

The tests and demo notebook have been updated accordingly.

A function smartg is kept for backward compatibility.
It may be removed in the future.

The function smartg_thr is removed (not so useful).