Optimization
Implement total photon accumulation in the GPU using gpuarray.
=> avoids getting GPU memory to central memory at each kernel
=> much faster accumulation
The total photon arrays are memcpied at the end.

=> Large speed improvement when lealing with large number of wavelengths.

Caveat: gpuarrays don't support broadcasting. Thus implementation
of stdev calculation requires particular implementation.
stdev calculation is still done on the CPU but leads to large slowdown.
It is then set to false by default, and a warning is raised on use.